strtoupper to mb_strtoupper
git-svn-id: https://svn.code.sf.net/p/web-erp/code/trunk@4622 09516a42-71f9-43be-811f-9b821cf0d246

+++ AccountGroups.php
@@ -262,7 +262,7 @@ or deletion of the records*/
 			<td>' . $PandLText . '</td>
 			<td>' . $myrow[4] . '</td>';
 		echo '<td><a href="' . $_SERVER['PHP_SELF'] . '?SelectedAccountGroup=' . htmlentities($myrow[0], ENT_QUOTES,'UTF-8') . '">' . _('Edit') . '</a></td>';
-		echo '<td><a href="' . $_SERVER['PHP_SELF'] . '?SelectedAccountGroup=' . htmlentities($myrow[0], ENT_QUOTES,'UTF-8') . '&amp;delete=1">' . _('Delete') .'</a></td></tr>';
+		echo '<td><a href="' . $_SERVER['PHP_SELF'] . '?SelectedAccountGroup=' . htmlentities($myrow[0], ENT_QUOTES,'UTF-8') . '&amp;delete=1" onclick="return confirm(\'' . _('Are you sure you wish to delete this account group?') . '\');">' . _('Delete') .'</a></td></tr>';
 
 	} //END WHILE LIST LOOP
 	echo '</table>';
